Fixed keyword and location searching, hide location, able to save website and address information from profile to the database
1. Like_me_matches.tpl & dynamic_bloc_user.tpl & users_list_block.tpl &
visits.tpl – since we remove location, make sure everything should point
to address.
2. Users.php – This is the file where we can pass data into the
database. Search function is inside. I edited the search so that for any
input string, we can look for company’s name that match with the string.
I also added two lines (on 387&388) to pass website and address into the
database when user makes a change on their account (profile &
registration)
3. Register.tpl – I added the address there, but still I keep it in hide
mode because I think it is optional to input address when creating an
account, users can add the address later.
I also hide the location in the register page.
4. Search.html – for the location searching, users were asked to select
the location that’s in the list, and it is not not flexible because
users can only search city, state, and country.
I made it so that users can search with any string and we will provide
the companies that their addresses have the given string.
5. View_my_profile.tpl & view_users_profile – hide the user look for
type and location attributes.