Commented variables like website, facebook & Show bank update on account services & No featured users on top of some pages
1. Service_form.tpl & Helper_users_carousel.tpl & list.tpl: if you see
the text or body element is not centered, please do so by adding this
line to the file, <div class="content"> Body </div>.
2. Users.php & users_model.php: commented those variables (like website,
facebook) that we do not use at this moment.
3.  Users-map.js: initialized the coordination (0,0) if no address is
inputted.
4. Account.tpl: uncommented the payment section which currently allows
users to update their account by inserting money.
5. Account_menu.tpl: change “History” to “Payment History”.
6. Dynamic_block_users.tpl: no age should be shown.
7. Helper_featured_carousel.tpl: featured users block should not be
shown on top of some pages like search, login, register pages, so I
deleted the line of code.